separated compose and mask functions
