+++ src/draw/sdl/lv_draw_sdl_rect.c
@@ -18,6 +18,7 @@
 #include "../../core/lv_refr.h"
 #include "lv_draw_sdl_utils.h"
 #include "lv_draw_sdl_texture_cache.h"
+#include "lv_draw_sdl_composite.h"
 #include "lv_draw_sdl_mask.h"
 #include "lv_draw_sdl_stack_blur.h"
 
@@ -141,7 +142,7 @@ void lv_draw_sdl_draw_rect(lv_draw_ctx_t * draw_ctx, const lv_draw_rect_dsc_t * 
     }
     /* Coords will be translated so coords will start at (0,0) */
     lv_area_t t_coords = *coords, t_clip = *clip, apply_area, t_area;
-    bool has_mask = lv_draw_sdl_mask_begin(ctx, coords, clip, &extension, dsc->blend_mode, &t_coords, &t_clip,
+    bool has_mask = lv_draw_sdl_composite_begin(ctx, coords, clip, &extension, dsc->blend_mode, &t_coords, &t_clip,
                                            &apply_area);
     bool has_content = _lv_area_intersect(&t_area, &t_coords, &t_clip);
 
@@ -156,7 +157,7 @@ void lv_draw_sdl_draw_rect(lv_draw_ctx_t * draw_ctx, const lv_draw_rect_dsc_t * 
     }
     draw_outline(ctx, &t_coords, &t_clip, dsc);
 
-    lv_draw_sdl_mask_end(ctx, has_mask, &apply_area, dsc->blend_mode);
+    lv_draw_sdl_composite_end(ctx, has_mask, &apply_area, dsc->blend_mode);
 }
 
 /**********************
